Add parsing for copyright dates
For the purpose of this application, the fact that a date is a
copyright date is irrelevant. It is a type of date that is
unfortunately often recorded as part of a date. There are
endless types of dates, and the type of date has no bearing on how the
date should be parsed. We basically just ignore copyright indicators.